build: Move public includes to their own directory
This makes it easy for in-tree consumers (e.g. add_subdirectory or
FetchContent) and consumers post-install (e.g. FindPackage) to have
the same include paths.
