+++ src/builder.h
@@ -21,7 +21,7 @@
 
 #include "common.h"
 #include "sentencepiece_model.pb.h"
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "third_party/absl/strings/string_view.h"
 
 namespace sentencepiece {
