+++ src/char_model_trainer_test.cc
@@ -17,7 +17,7 @@
 
 #include "char_model_trainer.h"
 #include "filesystem.h"
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "testharness.h"
 #include "third_party/absl/strings/str_cat.h"
 #include "third_party/absl/strings/str_join.h"
