+++ src/normalizer.h
@@ -24,7 +24,7 @@
 #include "common.h"
 #include "util.h"
 #include "sentencepiece_model.pb.h"
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "third_party/absl/strings/string_view.h"
 #include "third_party/darts_clone/darts.h"
