+++ src/pretokenizer_for_training.h
@@ -20,7 +20,7 @@
 
 #include "common.h"
 #include "sentencepiece.pb.h"
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "third_party/absl/strings/string_view.h"
 
 namespace sentencepiece {
