+++ src/spec_parser.h
@@ -18,7 +18,7 @@
 #include <string>
 #include <vector>
 
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "third_party/absl/strings/ascii.h"
 #include "third_party/absl/strings/str_split.h"
 #include "util.h"
