+++ src/spm_encode_main.cc
@@ -20,7 +20,7 @@
 #include "filesystem.h"
 #include "init.h"
 #include "sentencepiece.pb.h"
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "third_party/absl/container/flat_hash_map.h"
 #include "third_party/absl/flags/flag.h"
 #include "third_party/absl/strings/str_cat.h"
