+++ src/spm_train_main.cc
@@ -17,7 +17,7 @@
 #include "filesystem.h"
 #include "init.h"
 #include "sentencepiece_model.pb.h"
-#include "sentencepiece_trainer.h"
+#include <sentencepiece-browsermt/sentencepiece_trainer.h>
 #include "third_party/absl/flags/flag.h"
 #include "third_party/absl/strings/ascii.h"
 #include "third_party/absl/strings/str_join.h"
