+++ src/trainer_interface.cc
@@ -23,8 +23,8 @@
 #include "model_factory.h"
 #include "model_interface.h"
 #include "normalizer.h"
-#include "sentencepiece_processor.h"
-#include "sentencepiece_trainer.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
+#include <sentencepiece-browsermt/sentencepiece_trainer.h>
 #include "third_party/absl/container/flat_hash_map.h"
 #include "third_party/absl/memory/memory.h"
 #include "third_party/absl/strings/numbers.h"
