+++ src/trainer_interface.h
@@ -25,8 +25,8 @@
 #include "common.h"
 #include "filesystem.h"
 #include "sentencepiece_model.pb.h"
-#include "sentencepiece_processor.h"
-#include "sentencepiece_trainer.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
+#include <sentencepiece-browsermt/sentencepiece_trainer.h>
 #include "third_party/absl/container/flat_hash_map.h"
 #include "util.h"
