+++ src/util.h
@@ -29,7 +29,7 @@
 #include <vector>
 
 #include "common.h"
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "third_party/absl/strings/string_view.h"
 
 #ifdef SPM_NO_THREADLOCAL
