+++ src/word_model_trainer_test.cc
@@ -16,7 +16,7 @@
 #include <vector>
 
 #include "filesystem.h"
-#include "sentencepiece_processor.h"
+#include <sentencepiece-browsermt/sentencepiece_processor.h>
 #include "testharness.h"
 #include "third_party/absl/strings/str_cat.h"
 #include "third_party/absl/strings/str_join.h"
